backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiofiles
alembic
python-jose[cryptography]
bcrypt
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
python-dotenv
httpx
requests
auth0-python
auth0-fastapi-api
google-adk
google-genai
websockets
psycopg2-binary
python-dateutil
google-cloud-storage
